+++ logic/profession.lua
@@ -89,9 +89,7 @@ MTSL_LOGIC_PROFESSION = {
                     skill_passed_filter = false
                 end
                 -- Check availability in expansion
-                local used_expansion = v.expansion or 1
-                if not v.expansion then print(profession_name .. ' - ' .. v.id) end
-                if skill_passed_filter == true and MTSL_TOOLS:ListContainsNumber(expansions, 100 * used_expansion + v.phase) == false then
+                if skill_passed_filter == true and MTSL_TOOLS:ListContainsNumber(expansions, 100 * v.expansion + v.phase) == false then
                     skill_passed_filter = false
                 end
                 -- Check availability in zone