+++ ui/global_variables.lua
@@ -7,7 +7,7 @@ MTSLUI_ADDON = {
     AUTHOR = "Thumbkin",
     NAME = "Missing TradeSkills List (TBC)",
     VERSION = "2.5.06",
-    PATCH_TEXT = '* All professions except Jewelcrafting have all the correct skills, items and quests added\n' ..
+    PATCH_TEXT = '* All professions except Jewelcrafting & Leatherworking have all the correct skills, items and quests added\n' ..
                  '* Fixed data:\n' ..
                  '   * Set the correct data for each NPC\n' ..
                  '   * All items have the correct phase now, if they drop in specific zone(s), those zone(s) are added\n' ..