+++ VERSION_HISTORY.md
@@ -2,11 +2,12 @@
 
 ### v2.5.06 (June 14, 2021)
 
-* All professions except Jewelcrafting have all the skills & items added
+* All professions except Jewelcrafting & Leatherworking have all the skills & items added
 * Fixed data:
   * Set the correct data for each NPC
   * All items have the correct phase now, if they drop in specific zone(s), those zone(s) are added
   * Removed unused NPCs
+  * Removed unused or duplicate skills/items
   
 ### v2.5.05 (June 8, 2021)